Extract common function to a custom hook
Extracted the release tokens function to a custom hook so that it can be used
in any components.
